requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
xgboost>=1.5.0
matplotlib>=3.4.3
seaborn>=0.11.2
python-dotenv>=0.19.0
requests>=2.26.0
Pillow>=8.3.2
joblib>=1.1.0
groq>=0.4.0
tqdm>=4.62.3
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.2
mypy>=0.910
nltk>=3.6.5
opencv-python>=4.5.3
praw>=7.4.0
pytest-cov>=2.12.1
pre-commit>=2.15.0 